+++ Services/Administration/OJS.Services.Administration.Business/IParticipantsBusinessService.cs
@@ -1,0 +1,11 @@
+﻿namespace OJS.Services.Administration.Business;
+
+using SoftUni.Services.Infrastructure;
+using System.Threading.Tasks;
+
+public interface IParticipantsBusinessService : IService
+{
+    Task UpdateTotalScoreSnapshotOfParticipants();
+
+    Task RemoveParticipantMultipleScores();
+}
