+++ Services/Administration/OJS.Services.Administration.Data/IParticipantsDataService.cs
@@ -38,5 +38,7 @@ namespace OJS.Services.Administration.Data
         Task Delete(IEnumerable<Participant> participants);
 
         Task InvalidateByContestAndIsOfficial(int contestId, bool isOfficial);
+
+        Task UpdateTotalScoreSnapshot();
     }
 }
