+++ Services/UI/OJS.Services.Ui.Business/IParticipantsBusinessService.cs
@@ -40,9 +40,5 @@ namespace OJS.Services.Ui.Business
         Task<int> GetPracticeParticipantsCount(int contestId);
 
         Task<int> GetCompeteParticipantsCount(int contestId);
-
-        Task UpdateTotalScoreSnapshotOfParticipants();
-
-        Task RemoveParticipantMultipleScores();
     }
 }
