+++ Services/UI/OJS.Services.Ui.Data/IParticipantsDataService.cs
@@ -42,7 +42,5 @@ namespace OJS.Services.Ui.Data
         Task Delete(IEnumerable<Participant> participants);
 
         Task InvalidateByContestAndIsOfficial(int contestId, bool isOfficial);
-
-        Task UpdateTotalScoreSnapshot();
     }
 }
